requirements.txt
# RAG 시스템 전체 단계 통합 Requirements
# 1단계(PDF 파싱) ~ 6단계(LLM 통합) 통합 버전

# ==========================================
# PDF 처리
# ==========================================
pymupdf==1.23.8
camelot-py[cv]>=0.11.0

# ==========================================
# 데이터 처리
# ==========================================
pandas>=2.2.2
numpy==1.26.2

# ==========================================
# 시각화
# ==========================================
matplotlib==3.8.2
seaborn==0.13.0
pillow==10.1.0

# ==========================================
# 웹 앱 (Streamlit)
# ==========================================
streamlit>=1.28.0

# ==========================================
# OpenAI API
# ==========================================
openai>=1.12.0

# ==========================================
# 토큰 계산 (청킹용)
# ==========================================
tiktoken>=0.5.0

# ==========================================
# 벡터 검색 (FAISS)
# ==========================================
faiss-cpu>=1.7.4

# ==========================================
# 키워드 검색 (BM25)
# ==========================================
rank-bm25>=0.2.2

# ==========================================
# 머신러닝
# ==========================================
scikit-learn>=1.3.0

# ==========================================
# Jupyter 노트북
# ==========================================
jupyter==1.0.0
ipykernel==6.27.1
ipython>=8.0.0

# ==========================================
# 유틸리티
# ==========================================
python-dotenv==1.0.0